+++ src/SecucardConnect/Product/Payment/Model/TransferAccount.php
@@ -1,0 +1,32 @@
+<?php
+
+namespace SecucardConnect\Product\Payment\Model;
+
+class TransferAccount
+{
+    /**
+     * @var string
+     */
+    public $account_owner;
+
+    /**
+     * @var string
+     */
+    public $accountnumber;
+
+    /**
+     * @var string
+     */
+    public $iban;
+
+    /**
+     * @var string
+     */
+    public $bic;
+
+    /**
+     * @var string
+     */
+    public $bankcode;
+
+}